CodeGen CLI and Common library.

+++ cli/debug.cc
@@ -98,7 +98,7 @@ void DebugPrintNode(const Node* node) {
     }
 
     // Print child nodes.
-    for (int i = 0; i < node->child_count(); i++) {
+    for (size_t i = 0; i < node->child_count(); i++) {
         DebugPrintNode(node->child(i));
     }
 }
